moves SubsetDependencyMapper to pytato.transform
